requirements.txt
alembic==0.9.6
aniso8601==1.3.0
appnope==0.1.0
bcrypt==3.1.4
Blueprints==2.3.0.2
cffi==1.11.2
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.4.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
